+++ features/account/customer_account/changing_password.feature
@@ -6,13 +6,13 @@ Feature: Changing a customer password
 
     Background:
         Given the store operates on a single channel in "United States"
-        And there is a customer "Francis Underwood" identified by an email "francis@underwood.com" and a password "whitehouse"
+        And there is a customer "Francis Underwood" identified by an email "francis@underwood.com" and a password "sylius"
         And I am logged in as "francis@underwood.com"
 
-    @ui
+    @ui @shop_user_account_api
     Scenario: Changing my password
         When I want to change my password
-        And I change password from "whitehouse" to "blackhouse"
+        And I change password from "sylius" to "blackhouse"
         And I save my changes
         Then I should be notified that my password has been successfully changed
