+++ src/Sylius/Behat/Resources/config/services/api.xml
@@ -153,6 +153,11 @@
             <argument>shop</argument>
         </service>
 
+        <service id="sylius.behat.api_platform_client.shop.user" class="Sylius\Behat\Client\ApiPlatformClient" parent="sylius.behat.api_platform_client">
+            <argument>users</argument>
+            <argument>shop</argument>
+        </service>
+
         <service id="Sylius\Behat\Client\ResponseCheckerInterface" class="Sylius\Behat\Client\ResponseChecker">
             <argument type="service" id="test.client" />
         </service>
