+++ src/Sylius/Behat/Resources/config/suites/api/account/customer.yml
@@ -33,7 +33,7 @@ default:
                 - sylius.behat.context.setup.user
                 - sylius.behat.context.setup.zone
                     
-                - sylius.behat.context.api.shop.account
+                - sylius.behat.context.api.shop.customer
                 - sylius.behat.context.api.shop.order
                 - sylius.behat.context.api.shop.checkout
                 - sylius.behat.context.api.shop.login
