+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/Commands/ChangePasswordShopUser.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" ?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<resources xmlns="https://api-platform.com/schema/metadata"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="https://api-platform.com/schema/metadata https://api-platform.com/schema/metadata/metadata-2.5.xsd"
+>
+    <resource class="Sylius\Bundle\ApiBundle\Command\ChangePasswordShopUser" shortName="ChangePasswordShopUser">
+        <attribute name="route_prefix">shop</attribute>
+
+        <attribute name="output">false</attribute>
+
+        <attribute name="messenger">true</attribute>
+
+        <collectionOperations>
+            <collectionOperation name="post">
+                <attribute name="method">POST</attribute>
+                <attribute name="path">/change-password</attribute>
+                <attribute name="openapi_context">
+                    <attribute name="summary">Change user password</attribute>
+                </attribute>
+            </collectionOperation>
+        </collectionOperations>
+
+        <itemOperations />
+
+        <property name="password" required="true" />
+        <property name="oldPassword" required="true" />
+        <property name="confirmPassword" required="true" />
+    </resource>
+</resources>
