+++ src/Sylius/Bundle/ApiBundle/Resources/config/validation/ChangePasswordShopUser.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" ?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
+
+    <class name="Sylius\Bundle\ApiBundle\Command\ChangePasswordShopUser">
+        <property name="oldPassword">
+            <constraint name="Symfony\Component\Security\Core\Validator\Constraints\UserPassword">
+                <option name="message">sylius.user.plainPassword.wrong_current</option>
+                <option name="groups">sylius</option>
+            </constraint>
+            <constraint name="Length">
+                <option name="min">4</option>
+                <option name="minMessage">sylius.user.password.min</option>
+                <option name="max">254</option>
+                <option name="maxMessage">sylius.user.password.max</option>
+                <option name="groups">sylius</option>
+            </constraint>
+        </property>
+        <property name="password">
+            <constraint name="NotBlank">
+                <option name="groups">sylius</option>
+            </constraint>
+            <constraint name="Length">
+                <option name="min">4</option>
+                <option name="minMessage">sylius.user.password.min</option>
+                <option name="max">254</option>
+                <option name="maxMessage">sylius.user.password.max</option>
+                <option name="groups">sylius</option>
+            </constraint>
+        </property>
+    </class>
+</constraint-mapping>
